backend/requirements.txt
# AgriSense AI - Backend Dependencies
# Install with: pip install -r requirements.txt

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
scikit-learn>=1.3.0
joblib>=1.3.0
requests>=2.31.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0
Pillow>=10.0.0
numpy>=1.24.0
ultralytics>=8.0.0
torch>=2.0.0
